requirements.txt
instaloader
python-dotenv
flask
selenium
webdriver-manager
